Add an exception to source delete

+++ koku/api/provider/provider_manager.py
@@ -41,12 +41,21 @@ class ProviderManagerError(Exception):
         self.message = message
 
 
+class ProviderProcessingError(Exception):
+    """General Exception class for ProviderManager errors."""
+
+    def __init__(self, message):
+        """Set custom error message for ProviderManager errors."""
+        self.message = message
+
+
 class ProviderManager:
     """Provider Manager to manage operations related to backend providers."""
 
     def __init__(self, uuid):
         """Establish provider manager database objects."""
         self._uuid = uuid
+        self.date_helper = DateHelper()
         try:
             self.model = Provider.objects.get(uuid=self._uuid)
         except (ObjectDoesNotExist, ValidationError) as exc:
@@ -78,7 +87,7 @@ class ProviderManager:
         """Get current month data avaiability status."""
         return CostUsageReportManifest.objects.filter(
             provider=self._uuid,
-            billing_period_start_datetime=DateHelper().this_month_start,
+            billing_period_start_datetime=self.date_helper.this_month_start,
             manifest_completed_datetime__isnull=False,
         ).exists()
 
@@ -86,7 +95,7 @@ class ProviderManager:
         """Get current month data avaiability status."""
         return CostUsageReportManifest.objects.filter(
             provider=self._uuid,
-            billing_period_start_datetime=DateHelper().last_month_start,
+            billing_period_start_datetime=self.date_helper.last_month_start,
             manifest_completed_datetime__isnull=False,
         ).exists()
 
@@ -94,6 +103,15 @@ class ProviderManager:
         """Get  data avaiability status."""
         return CostUsageReportManifest.objects.filter(
             provider=self._uuid, manifest_completed_datetime__isnull=False
+        ).exists()
+
+    def get_is_provider_processing(self):
+        """Return a bool determining if the source is currently processing."""
+        return CostUsageReportManifest.objects.filter(
+            provider=self._uuid,
+            billing_period_start_datetime=self.date_helper.this_month_start,
+            manifest_creation_datetime__date=self.date_helper.today.date(),
+            manifest_completed_datetime__isnull=True,
         ).exists()
 
     def get_infrastructure_info(self):
@@ -217,6 +235,10 @@ class ProviderManager:
         if self.sources_model and not from_sources:
             err_msg = f"Provider {self._uuid} must be deleted via Sources Integration Service"
             raise ProviderManagerError(err_msg)
+
+        if from_sources and self.get_is_provider_processing():
+            err_msg = f"Provider {self._uuid} is currently being processed and must finish before delete."
+            raise ProviderProcessingError(err_msg)
 
         if self.is_removable_by_user(current_user):
             self.model.delete()
